pyproject.toml
[project]
name = "matriosha"
version = "2.0.0"
description = "Matriosha is a Python CLI for encrypted agent memory with verifiable integrity."
requires-python = ">=3.11,<3.14"
dependencies = [
  "typer>=0.12,<0.13",
  "rich>=13.7,<14",
  "cryptography>=42,<43",
  "argon2-cffi>=23.1,<24",
  "pynacl>=1.5,<2",
  "requests>=2.32,<3",
  "httpx>=0.27,<0.28",
  "supabase>=2.5,<3",
  "pydantic>=2.7,<3",
  "numpy>=1.26,<2",
  "platformdirs>=4.2,<5",
  "google-cloud-secret-manager>=2.16,<3",
  "pypdf>=4.2,<6",
  "python-docx>=1.1,<2",
  "openpyxl>=3.1,<4",
  "Pillow>=10.3,<12",
  "pytesseract>=0.3.10,<0.4",
]

[project.scripts]
matriosha = "matriosha.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/matriosha"]

[tool.hatch.build.targets.sdist]
include = [
  "src/matriosha",
  "pyproject.toml",
  "requirements.txt",
  "README.md",
]

[project.optional-dependencies]
cloud = [
  "google-cloud-secret-manager>=2.16,<3",
]
cli-ux = [
  "google-cloud-secret-manager>=2.16,<3",
  "qrcode[pil]>=7.4.2,<8",
]
tui = [
  "textual>=0.70,<1",
  "pytailwindcss>=0.0.4,<1",
]
embeddings = [
  "sentence-transformers>=3,<4",
]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "respx>=0.22,<0.23",
  "syrupy>=4.8,<5",
  "pexpect>=4.9,<5",
  "ruff>=0.5",
  "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -ra"
markers = [
  "integration: e2e scenarios",
  "adversarial: break-things and resilience scenarios",
  "managed: scenarios requiring real managed backend credentials"
]
